requirements.txt
Flask==2.0.1
requests==2.25.1
pyyaml>=6.0
azure-mgmt-compute==23.1.0
azure-mgmt-network==21.0.0
azure-identity==1.10.0
python-dotenv==0.19.2
pytest==7.0.1
scapy==2.4.5
cryptography==36.0.1
tensorflow==2.8.0
numpy==1.22.2
pandas==1.4.1
scikit-learn==1.2.2